Modularize codebase and bundle with parcel
