+++ src/manifest.json
@@ -8,11 +8,8 @@
   "content_scripts": [
     {
       "matches": ["https://github.com/*/projects/*"],
-      "js": [
-        "node_modules/webextension-polyfill/dist/browser-polyfill.js",
-        "projects.js"
-      ],
-      "css": ["projects.css"],
+      "js": ["index.js"],
+      "css": ["index.css"],
       "run_at": "document_end"
     }
   ],
